limit tiebreaker to 16 bits

+++ quickwit/quickwit-processing/src/processed_log.rs
@@ -118,7 +118,7 @@ pub struct ProcessedLog {
 
     pub id: String,
     pub discovery_timestamp: i64,
-    pub tiebreaker: i32,
+    pub tiebreaker: i16,
     pub ingest_size_in_bytes: usize,
 }
